pushing what I did so far
