+++ src/Kanbas/index.tsx
@@ -41,12 +41,13 @@ export default function Kanbas() {
                     <Route path="/Account/*" element={<Account/>}/>
                     <Route path={"/Dashboard"} element={
                         <ProtectedRoute><Dashboard
-                        courses={courses}
-                        course={course}
-                        setCourse={setCourse}
-                        addNewCourse={addNewCourse}
-                        deleteCourse={deleteCourse}
-                        updateCourse={updateCourse}/></ProtectedRoute>
+                            courses={courses}
+                            course={course}
+                            setCourse={setCourse}
+                            addNewCourse={addNewCourse}
+                            deleteCourse={deleteCourse}
+                            updateCourse={updateCourse}/>
+                        </ProtectedRoute>
                     }/>
                     <Route path="/Courses/:cid/*" element={<ProtectedRoute><Courses courses={courses}/></ProtectedRoute>} />
                     <Route path="/Calendar" element={<h1>Calendar</h1>} />
