+++ src/Kanbas/store.ts
@@ -2,11 +2,13 @@ import { configureStore } from "@reduxjs/toolkit";
 import modulesReducer from "./Courses/Modules/reducer";
 import accountReducer from "./Account/reducer";
 import assignmentsReducer from "./Courses/Assignments/reducer";
+import enrollementReducer from "./Dashboard/reducer"
 const store = configureStore({
     reducer: {
         modulesReducer,
         accountReducer,
         assignmentsReducer,
+        enrollementReducer,
     },
 });
 export default store;
